added errors to gameController

+++ server/src/controllers/game/gameController.js
@@ -3,6 +3,14 @@ import playerController from "../../controllers/player/playerController.js";
 import questionnaireModel from "../../models/questionnaire.js";
 import playerModel from "../../models/player.js"
 import { PlayerDoesNotExist } from "../../utils/errors/playerErrors.js"
+import {
+  GameDoesNotExist,
+  QuestionsNotProvided,
+  QuestionnaireNotProvided,
+  HostNotProvided
+} from "../../utils/errors/gameErrors.js";
+
+import { QuestionnaireDoesNotExist } from "../../utils/errors/questionnaireErrors.js";
 import { customAlphabet } from "nanoid";
 
 const getRandomCode = customAlphabet("ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789", 6);
@@ -11,22 +19,28 @@ async function getGameById(id) {
   const game = await gameModel.findById(id) //Get the game using the id
     .populate("players", "nickname") //add the players
     .populate("questionnaireId"); //add a questionnaire
+
+  if (!game) throw new GameDoesNotExist();
   return game;
 }
 
 async function createGame(host, questionnaireId) {
+
+  if(!host) throw new HostNotProvided();
+  if(!questionnaireId) throw new QuestionnaireNotProvided();
+  //Delete the games that are active of the host
+  //A host can only have one active game
   await gameModel.deleteMany({
     host: host,
     questionnaireId: questionnaireId
   });
-
+  //Create the new game
   const game = new gameModel({
     host: host,
     questionnaireId: questionnaireId,
     code: getRandomCode()
   });
   await game.save();
-
   return game;
 }
 
@@ -36,6 +50,7 @@ async function joinPlayer(nickname, gameId) { //meter _id
 
 async function startGame(gameId) {
   const game = await gameModel.findById(gameId);
+  if (!game) throw new GameDoesNotExist();
   game.state = "started";
   game.updatedAt = new Date();
   await game.save();
@@ -45,8 +60,11 @@ async function startGame(gameId) {
 
 async function nextQuestion(gameId) {
   const game = await gameModel.findById(gameId).populate("questionnaireId");
+  if (!game) throw new GameDoesNotExist();
+
   game.questionIndex++;
   const questions = game.questionnaireId.questions;
+  if(!questions) throw new QuestionsNotProvided();
 
   if (game.questionIndex >= questions.length) {
     game.state = "finished";
@@ -61,7 +79,9 @@ async function nextQuestion(gameId) {
 
 async function getQuestion(gameId) {
   const game = await gameModel.findById(gameId);
+  if (!game) throw new GameDoesNotExist();
   const questionnaire = await questionnaireModel.findById(game.questionnaireId);
+  if(!questionnaire) throw new QuestionnaireDoesNotExist();
   return questionnaire.questions[game.questionIndex];
 }
 
@@ -83,6 +103,5 @@ export default {
   startGame,
   nextQuestion,
   getQuestion,
-  removeGame,
   saveSocketIdToPlayer
 }
